+++ changes.md
@@ -12,5 +12,5 @@ Change Log
 * added the missing hashes of ExternalReference
 * remove description from ObservablesMaker
 * added custom (for custom properties x_...) to ObservablesMaker
-* update to scalastix-06, ie. can process Custom Objects
+* update to scalastix-0.6, ie. can process Custom Objects