+++ changes.md
@@ -12,5 +12,9 @@ Change Log
 * added the missing hashes of ExternalReference
 * remove description from ObservablesMaker
 * added custom (for custom properties x_...) to ObservablesMaker
-* update to scalastix-0.6, ie. can process Custom Objects
+* update to scalastix-0.7, for STIX-2.0 specs 
+* due to STIX-2.0 specs, 
+** modify object_refs as optional for Report
+** comment out the lang and confidence parameters.
 
+* updated scala, sbt, plugins